For #24780 - Part 4: Refactor AutofillSettingFragment to its own package
